SessionStore rename
