+++ app/src/common/shared/org/mozilla/vrbrowser/DataRepository.java
@@ -7,7 +7,7 @@ import androidx.lifecycle.LifecycleRegistry;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MediatorLiveData;
 
-import org.mozilla.vrbrowser.browser.engine.SessionManager;
+import org.mozilla.vrbrowser.browser.engine.SessionStore;
 import org.mozilla.vrbrowser.db.AppDatabase;
 import org.mozilla.vrbrowser.db.entity.BookmarkEntity;
 import org.mozilla.vrbrowser.model.Bookmark;
@@ -75,7 +75,7 @@ public class DataRepository implements LifecycleOwner {
                 return;
             }
             for (BookmarkEntity oldBookmark: oldBookmarks) {
-                SessionManager.get().getBookmarkStore().addBookmark(oldBookmark.getUrl(), oldBookmark.getTitle());
+                SessionStore.get().getBookmarkStore().addBookmark(oldBookmark.getUrl(), oldBookmark.getTitle());
             }
             mExecutors.diskIO().execute(() -> mDatabase.bookmarkDao().deleteAll());
         });
