+++ app/src/common/shared/org/mozilla/vrbrowser/ui/views/NavigationURLBar.java
@@ -31,9 +31,8 @@ import org.mozilla.geckoview.GeckoSessionSettings;
 import org.mozilla.vrbrowser.R;
 import org.mozilla.vrbrowser.audio.AudioEngine;
 import org.mozilla.vrbrowser.browser.BookmarksStore;
-import org.mozilla.vrbrowser.browser.engine.SessionManager;
 import org.mozilla.vrbrowser.browser.engine.SessionStore;
-import org.mozilla.vrbrowser.browser.SettingsStore;
+import org.mozilla.vrbrowser.browser.engine.SessionStack;
 import org.mozilla.vrbrowser.search.SearchEngineWrapper;
 import org.mozilla.vrbrowser.telemetry.TelemetryWrapper;
 import org.mozilla.vrbrowser.ui.widgets.WidgetPlacement;
@@ -73,7 +72,7 @@ public class NavigationURLBar extends FrameLayout {
     private boolean mBookmarkEnabled = true;
     private boolean mIsContextButtonsEnabled = true;
     private UIThreadExecutor mUIThreadExecutor = new UIThreadExecutor();
-    private SessionStore mSessionStore;
+    private SessionStack mSessionStack;
 
     private Unit domainAutocompleteFilter(String text) {
         if (mURL != null) {
@@ -105,7 +104,7 @@ public class NavigationURLBar extends FrameLayout {
     private void initialize(Context aContext) {
         mAudio = AudioEngine.fromContext(aContext);
 
-        mSessionStore = SessionManager.get().getActiveStore();
+        mSessionStack = SessionStore.get().getActiveStore();
 
         // Inflate this data binding layout
         inflate(aContext, R.layout.navigation_url, this);
@@ -155,7 +154,7 @@ public class NavigationURLBar extends FrameLayout {
         mUAModeButton = findViewById(R.id.uaModeButton);
         mUAModeButton.setTag(R.string.view_id_tag, R.id.uaModeButton);
         mUAModeButton.setOnClickListener(mUAModeListener);
-        setUAMode(mSessionStore.getUaMode());
+        setUAMode(mSessionStack.getUaMode());
 
         mURLLeftContainer = findViewById(R.id.urlLeftContainer);
         mInsecureIcon = findViewById(R.id.insecureIcon);
@@ -181,9 +180,9 @@ public class NavigationURLBar extends FrameLayout {
         syncViews();
     }
 
-    public void setSessionStore(SessionStore sessionStore) {
-        mSessionStore = sessionStore;
-        setUAMode(mSessionStore.getUaMode());
+    public void setSessionStack(SessionStack sessionStack) {
+        mSessionStack = sessionStack;
+        setUAMode(mSessionStack.getUaMode());
     }
 
     public void onPause() {
@@ -242,14 +241,14 @@ public class NavigationURLBar extends FrameLayout {
             mAudio.playSound(AudioEngine.Sound.CLICK);
         }
 
-        String url = mSessionStore.getCurrentUri();
+        String url = mSessionStack.getCurrentUri();
         if (StringUtils.isEmpty(url)) {
             return;
         }
-        BookmarksStore bookmarkStore = SessionManager.get().getBookmarkStore();
+        BookmarksStore bookmarkStore = SessionStore.get().getBookmarkStore();
         bookmarkStore.isBookmarked(url).thenAcceptAsync(bookmarked -> {
             if (!bookmarked) {
-                bookmarkStore.addBookmark(url, mSessionStore.getCurrentTitle());
+                bookmarkStore.addBookmark(url, mSessionStack.getCurrentTitle());
                 setBookmarked(true);
             } else {
                 // Delete
@@ -281,7 +280,7 @@ public class NavigationURLBar extends FrameLayout {
         if (StringUtils.isEmpty(aURL)) {
             setBookmarked(false);
         } else {
-            SessionManager.get().getBookmarkStore().isBookmarked(aURL).thenAcceptAsync(this::setBookmarked, mUIThreadExecutor);
+            SessionStore.get().getBookmarkStore().isBookmarked(aURL).thenAcceptAsync(this::setBookmarked, mUIThreadExecutor);
         }
 
         int index = -1;
@@ -294,9 +293,9 @@ public class NavigationURLBar extends FrameLayout {
             }
             if (aURL.startsWith("jar:"))
                 return;
-            else if (aURL.startsWith("resource:") || mSessionStore.isHomeUri(aURL))
+            else if (aURL.startsWith("resource:") || mSessionStack.isHomeUri(aURL))
                 aURL = "";
-            else if (aURL.startsWith("data:") && mSessionStore.isPrivateMode())
+            else if (aURL.startsWith("data:") && mSessionStack.isPrivateMode())
                 aURL = "";
             else
                 index = aURL.indexOf("://");
@@ -457,9 +456,9 @@ public class NavigationURLBar extends FrameLayout {
             TelemetryWrapper.urlBarEvent(false);
         }
 
-        if (mSessionStore.getCurrentUri() != url) {
-            SessionManager.get().getHistoryStore().addHistory(url, VisitType.TYPED);
-            mSessionStore.loadUri(url);
+        if (mSessionStack.getCurrentUri() != url) {
+            SessionStore.get().getHistoryStore().addHistory(url, VisitType.TYPED);
+            mSessionStack.loadUri(url);
 
             if (mDelegate != null) {
                 mDelegate.onHideSearchPopup();
@@ -501,14 +500,14 @@ public class NavigationURLBar extends FrameLayout {
 
         view.requestFocusFromTouch();
 
-        int uaMode = mSessionStore.getUaMode();
+        int uaMode = mSessionStack.getUaMode();
         if (uaMode == GeckoSessionSettings.USER_AGENT_MODE_DESKTOP) {
             setUAMode(GeckoSessionSettings.USER_AGENT_MODE_VR);
-            mSessionStore.setUaMode(GeckoSessionSettings.USER_AGENT_MODE_VR);
+            mSessionStack.setUaMode(GeckoSessionSettings.USER_AGENT_MODE_VR);
 
         }else {
             setUAMode(GeckoSessionSettings.USER_AGENT_MODE_DESKTOP);
-            mSessionStore.setUaMode(GeckoSessionSettings.USER_AGENT_MODE_DESKTOP);
+            mSessionStack.setUaMode(GeckoSessionSettings.USER_AGENT_MODE_DESKTOP);
         }
 
         TelemetryWrapper.voiceInputEvent();
