+++ app/src/common/shared/org/mozilla/vrbrowser/ui/widgets/settings/EnvironmentOptionsView.java
@@ -6,12 +6,11 @@
 package org.mozilla.vrbrowser.ui.widgets.settings;
 
 import android.content.Context;
-import android.widget.ScrollView;
 
 import org.mozilla.vrbrowser.R;
 import org.mozilla.vrbrowser.audio.AudioEngine;
 import org.mozilla.vrbrowser.browser.SettingsStore;
-import org.mozilla.vrbrowser.browser.engine.SessionManager;
+import org.mozilla.vrbrowser.browser.engine.SessionStore;
 import org.mozilla.vrbrowser.ui.views.UIButton;
 import org.mozilla.vrbrowser.ui.views.settings.ButtonSetting;
 import org.mozilla.vrbrowser.ui.views.settings.ImageRadioGroupSetting;
@@ -56,7 +55,7 @@ class EnvironmentOptionsView extends SettingsView {
             if (mAudio != null) {
                 mAudio.playSound(AudioEngine.Sound.CLICK);
             }
-            SessionManager.get().getActiveStore().loadUri(getContext().getString(R.string.environment_override_help_url));
+            SessionStore.get().getActiveStore().loadUri(getContext().getString(R.string.environment_override_help_url));
             exitWholeSettings();
         });
